pyproject.toml
[tool.poetry]
name = "discord-reactor"
version = "0.1.0"
description = ""
authors = ["Shogo Takata <pineapplehunter.daniel@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
"discord.py" = "^1.6.0"

[tool.poetry.dev-dependencies]
autopep8 = "^1.5.5"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
reactor = "discord_reactor.main:main"